@@ -6,23 +6,24 @@
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
     "axios": "^0.26.1",
-    "framer-motion": "^4.1.17",
+    "framer-motion": "^6.0.0",
     "graphql": "^16.3.0",
     "graphql-request": "^4.2.0",
     "lodash": "^4.17.21",
     "mobx": "^6.1.8",
     "mobx-react-lite": "^3.2.0",
     "moment": "^2.29.1",
-    "node-sass": "^5.0.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "node-sass": "7.0.1",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "react-hook-form": "^7.30.0",
     "react-intersection-observer": "^8.31.0",
     "react-query": "^3.34.20",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
     "react-toastify": "^7.0.4",
-    "react-transition-group": "1.x",
-    "styled-components": "^5.2.1",
+    "react-transition-group": "4.4.2",
+    "styled-components": "^5.3.5",
     "web-vitals": "^1.0.1",
     "workbox-background-sync": "^6.1.2",
     "workbox-broadcast-update": "^6.1.2",
@@ -38,8 +39,10 @@
     "workbox-streams": "^6.1.2"
   },
   "scripts": {
-    "start": "yarn gen-gql && react-scripts start",
+    "start": " react-scripts start",
     "build": "react-scripts build",
+    "start-prod": "yarn build && serve -s build",
+    "build:dev": "env-cmd -f .env.development react-scripts build && serve -s build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "gen-gql": "graphql-codegen"
@@ -63,27 +66,33 @@
     ]
   },
   "devDependencies": {
-    "@graphql-codegen/cli": "2.6.2",
-    "@graphql-codegen/fragment-matcher": "3.2.1",
-    "@graphql-codegen/introspection": "2.1.1",
-    "@graphql-codegen/schema-ast": "^2.4.1",
-    "@graphql-codegen/typescript": "2.4.8",
-    "@graphql-codegen/typescript-operations": "2.3.5",
+    "@babel/core": "^7.17.10",
+    "@babel/plugin-syntax-flow": "^7.16.7",
+    "@babel/plugin-transform-react-jsx": "^7.17.3",
+    "@babel/plugin-transform-react-jsx-source": "^7.16.7",
+    "@graphql-codegen/cli": "^2.6.2",
+    "@graphql-codegen/typescript": "^2.4.8",
+    "@graphql-codegen/typescript-operations": "^2.3.5",
     "@graphql-codegen/typescript-react-query": "^3.5.9",
-    "@graphql-codegen/typescript-resolvers": "2.6.1",
+    "@material-ui/types": "^5.1.0",
+    "@testing-library/dom": "^8.13.0",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "@types/axios": "^0.14.0",
     "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/react-transition-group": "1.x",
+    "@types/node": "^17.0.31",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "@types/react-router-dom": "^5.3.1",
+    "@types/react-transition-group": "4.4.4",
     "@types/styled-components": "^5.1.7",
-    "typescript": "^4.1.2",
-    "@graphql-codegen/typescript-graphql-files-modules": "2.1.1",
-    "@graphql-codegen/typescript-document-nodes": "2.2.8"
-  }
+    "browserslist": "^4.20.3",
+    "env-cmd": "^10.1.0",
+    "postcss": "^8.4.13",
+    "postcss-normalize": "^10.0.1",
+    "react-is": "^18.1.0",
+    "typescript": "^4.6.4"
+  },
+  "peerDependencies": {}
 }
