@@ -4,11 +4,11 @@ import "./App.css";
 import { Home } from "../home";
 import { TodoListOverview } from "../todoList";
 import Header from "../home/pages/header";
-import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 
-const App: React.FC = () => {
+const App: React.FC<React.PropsWithChildren<unknown>> = () => {
   return (
     <>
       <ToastContainer
@@ -21,18 +21,15 @@ const App: React.FC = () => {
         draggable
       />
       <Router>
-        <Header  />
-        <Switch>
-          <Route exact path="/" component={Home} />
-
-          <Route path="/todo-app">
-            <TodoListOverview />
-          </Route>
+        <Header />
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/todo-app" element={<TodoListOverview />} />
           {/* 
           <Route path="/dashboard">
             <Dashboard />
           </Route> */}
-        </Switch>
+        </Routes>
       </Router>
     </>
   );
