@@ -3,7 +3,7 @@ import React from "react";
 import MainPage from "./pages/mainPage";
 import Content from "./pages/content";
 import Footer from "./pages/footer";
-import { RouteComponentProps } from "react-router";
+import { useNavigate } from "react-router-dom";
 
 // import { motion, useAnimation, useMotionValue } from "framer-motion";
 // import styled from "styled-components";
@@ -12,16 +12,13 @@ import { RouteComponentProps } from "react-router";
 //   open: boolean;
 //   onFormSubmit: (value: string) => void;
 // }
-const Home: React.FC<{} & RouteComponentProps<{}>> = ({
-  match: { path },
-  history,
-}) => {
+const Home: React.FC<React.PropsWithChildren<{}>> = () => {
   // const [value, setValue] = useState("");
-
+  const navigate = useNavigate();
   return (
     <>
       <div style={{ overflow: "hidden" }}>
-        <MainPage onStartApp={() => history.push("todo-app")} />
+        <MainPage onStartApp={() => navigate("todo-app")} />
       </div>
       <Content />
       <Footer />
